+++ lib/screens_shared/settings.dart
@@ -12,10 +12,12 @@ import 'package:jne_household_app/i18n/i18n.dart';
 import 'package:jne_household_app/screens_shared/remote_database.dart';
 import 'package:jne_household_app/widgets_shared/dialogs/adaptive_alert_dialog.dart';
 import 'package:jne_household_app/widgets_shared/dialogs/debug_import_file_dialog.dart';
+import 'package:jne_household_app/widgets_shared/dialogs/report_issue_dialog.dart';
 import 'package:jne_household_app/widgets_shared/settings/bank_account.dart';
 import 'package:jne_household_app/widgets_shared/settings/export_import.dart';
 import 'package:jne_household_app/widgets_shared/settings/language.dart';
 import 'package:local_auth/local_auth.dart';
+import 'package:mail_sender/mail_sender.dart';
 import 'package:path/path.dart' as p;
 import 'package:path_provider/path_provider.dart';
 import 'package:provider/provider.dart';
@@ -276,10 +278,29 @@ class SettingsScreenState extends State<SettingsScreen> {
               },
               child: Text(I18n.translate("showLogs"))
             ),
+            const SizedBox(height: 8,),
+            ElevatedButton(
+              style: btnNeutralStyle,
+              onPressed: () async {
+                List<String> data = await showReportIssueDialog(context, budgetState);
+                if (data.length == 2) {
+                  if (Platform.isAndroid || Platform.isIOS) {
+                    final mailSenderPlugin = MailSender();
+                    mailSenderPlugin.sendMail(
+                      recipient: [reportEmail],
+                      subject: data[0],
+                      body: data[1]
+                    );
+                  }
+                }
+              },
+              child: Text(I18n.translate("report")),
+            ),
             if (kDebugMode)
-            ElevatedButton(onPressed: () => budgetState.updateIsPro(!budgetState.isPro), child: Text("DEBUG: Toggle pro\nCurrent: ${budgetState.isPro}")),
-            if (kDebugMode)
-            ElevatedButton(onPressed: () => {showDummyImportDialog(context, budgetState)}, child: const Text("DEBUG: Fast import pbstate lang file"))
+            ...[
+              ElevatedButton(onPressed: () => budgetState.updateIsPro(!budgetState.isPro), child: Text("DEBUG: Toggle pro\nCurrent: ${budgetState.isPro}")),
+              ElevatedButton(onPressed: () => {showDummyImportDialog(context, budgetState)}, child: const Text("DEBUG: Fast import pbstate lang file"))
+            ]
           ]
         )
       ),