+++ changelog.md
@@ -4,7 +4,6 @@
 
 ### 3.2.0
 
-- report issue button to submit error reports from within the app
 - custom background gradients
 - notification service
 - "real" credit card bank account
@@ -22,6 +21,10 @@ Integration of mobile payment apps is impossible due to privacy reasons maybe pu
 - add connection to iCloud
 
 ## Changelog
+
+### 3.2.0 in doing
+
+- report issue button to submit error reports from within the app
 
 ### 3.1.2 bundle 29 (27.08.2025)